+++ snovault/tests/root.py
@@ -67,7 +67,8 @@ class TestRoot(Root):
     def __acl__(self):
         acl = acl_from_settings(self.registry.settings) + [
             (Allow, Everyone, ['list', 'search']),
-            (Allow, 'group.admin', ALL_PERMISSIONS)
+            (Allow, 'group.admin', ALL_PERMISSIONS),
+            (Allow, 'remoteuser.EMBED', 'restricted_fields'),
         ] + [(Allow, 'remoteuser.INDEXER', ['view', 'view_raw', 'list', 'index']),
         (Allow, 'remoteuser.EMBED', ['view', 'view_raw', 'expand']),
         (Allow, Everyone, ['visible_for_edit'])]